requirements.txt
discord.py
apscheduler
aiosqlite